big step semantics (kind of)
K doesn't support big step semantics, but we can fake it in the easy
case. In the slightly harder case in which there is more than one rule
for the same construct (with if-then-else for example), a hybrid between
big step and small step is used.

I originally wanted to keep both the small step and the big step
operational semantics as a demonstration, but it turns out those two
cannot share the same syntax file because the construct need different
strictness annotations. So I will only keep the big step semantics.